+++ src/page/settings/components/section.jsx
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 
 const Section = styled.div`
   max-width: 680px;
-  background-color: #404040;
+  background-color: #2b2b2b;
   margin: 10px;
   padding: 10px;
   h2 {
@@ -12,6 +12,10 @@ const Section = styled.div`
     font-size: 11pt;
     margin-top: 5px;
     margin-bottom: 5px;
+  }
+
+  & > div {
+    margin-top: 5px;
   }
 `;