+++ src/page/settings/pages/defaults.jsx
@@ -1,24 +1,9 @@
 import React, { useReducer } from 'react';
-import styled from 'styled-components';
 import InputContainer from '../../editprofile/components/inputcontainer';
 import Checkbox from '../../../component/checkbox/checkbox';
 import Detail from '../../../component/detail/detail';
 import SettingsManager from '../../../manager/settingsManager';
-
-const Panel = styled.div`
-  background-color: #2b2b2b;
-  padding: 10px;
-  width: 380px;
-  margin-bottom: 5px;
-
-  h3 {
-    margin: 0;
-  }
-
-  & > div {
-    margin-top: 5px;
-  }
-`;
+import Section from '../components/section';
 
 export default function Defaults() {
   const [, forceUpdate] = useReducer(x => x + 1, 0);
@@ -34,7 +19,7 @@ export default function Defaults() {
       <p>
         The following are default settings that are used for <b>new</b> profiles only
       </p>
-      <Panel>
+      <Section>
         <h3>Sync Options</h3>
         <InputContainer>
           <Checkbox
@@ -60,7 +45,7 @@ export default function Defaults() {
           />
           <Detail>Sync in-game server list</Detail>
         </InputContainer>
-      </Panel>
+      </Section>
     </>
   );
 }